[import] add management of lang, genres & assets of the books we import
Instead of storing the full Genres titles every time (they van be quite long), we compute a hashid for each of them - after having first generated a adler32 sum from the Genre title. Hacky? Yes!

We also maintain "genres:stats:books_by_lang:{genre_hash}" Redis hashes, where each key is a lang and each value the number of books we have for this genre and this lang. Will probably be useful later on.

It now takes 20s to import 2000 books, but that's still ok since we don't optimise for performance at all :-)